# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Orbit Fundraising Intelligence API
  version: 1.0.0
  description: |
    Backend API for the Orbit autonomous fundraising platform.
    Serves the dashboard frontend and orchestrates CRM sync, AI agents, and webhooks.
servers:
  - url: http://localhost:3001/api/v1
    description: Local development
  - url: https://api.orbitfundraising.com/api/v1
    description: Production
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Campaign:
      type: object
      properties:
        id:           { type: string }
        name:         { type: string }
        status:       { type: string, enum: [draft, live, paused, completed] }
        assignedAgent:{ type: string }
        goal:         { type: number }
        raised:       { type: number }
        donorCount:   { type: integer }
        channel:      { type: string }
        startDate:    { type: string, format: date }
        endDate:      { type: string, format: date }
...
paths:
...
  /campaigns:
    get:
      summary: List campaigns
      tags: [Campaigns]
      responses:
        '200':
          description: Campaign list

    post:
      summary: Create campaign
      tags: [Campaigns]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Campaign' }
      responses:
        '201':
          description: Created
  /campaigns/{id}:
    get:
      summary: Get campaign
      tags: [Campaigns]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Campaign with full stats

    patch:
      summary: Update campaign
      tags: [Campaigns]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Campaign' }
      responses:
        '200':
          description: Updated

  # ── INTEGRATIONS ──────────────────────────────────────────────────────────────
...